requirements.txt
requests
ffmpeg-python
aiohttp
ffmpeg
hachoir
heroku3
wget
speedtest-cli
Pillow
pyromod
psutil
TgCrypto
motor
dnspython
python-dotenv
git+https://github.com/yt-dlp/yt-dlp
pyrogram
#https://github.com/Tinche/aiofiles/archive/master.zip
aiofiles
pyrogram
tgcrypto
motor
marshmallow
umongo
requests
psutil
pymongo[srv]
python-dotenv
pytz
requests
waybackpy
heroku3
pyrogram
tgcrypto
motor
marshmallow
umongo
requests
psutil
pymongo[srv]
python-dotenv
pytz
requests
waybackpy
heroku3
